+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 4.1.3 (Unreleased)
 
 - New config setting [locked_muc_domain](https://conversejs.org/docs/html/configuration.html#locked-muc-domain)
+- New config setting [show_client_info](https://conversejs.org/docs/html/configuration.html#show-client-info)
 - #1373: Re-add support for the [muc_domain](https://conversejs.org/docs/html/configuration.html#muc-domain) setting
 - #1400: When a chat message is just an emoji, enlarge the emoji
 - #1437: List of groupchats in modal doesn't scroll